Source Commit: Rearranged folders to make more sense. Also created an enemy with an idle animation and a functioning health bar that appears when the enemy is shot, disappearing after death or not being hit for a short period of time.

+++ Source/BoomerShooter/BaseCharacter.h
@@ -77,15 +77,15 @@ private:
 	int ActiveIndex = 0;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Health", meta = (AllowPrivateAccess = true))
-	int MaxHealth;
+	float MaxHealth;
 
 	UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Health", meta = (AllowPrivateAccess = true))
-	int Health;
+	float Health;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Mana", meta = (AllowPrivateAccess = true))
-	int MaxMana;
+	float MaxMana;
 
 	UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Mana", meta = (AllowPrivateAccess = true))
-	int Mana;
+	float Mana;
 
 };
